Move proper hat switch handling from ps4dshock to hgame
Generic "DirectInput" HID gamepads need this handling too.
