+++ hgame.h
@@ -31,6 +31,7 @@
 
 #include "hidmap.h"
 
+hidmap_cb_t	hgame_hat_switch_cb;
 hidmap_cb_t	hgame_dpad_cb;
 hidmap_cb_t	hgame_final_cb;
